Add typos and spelling errors.
it is actually spelt "Sce.PlayStation.Core.Graphics cannot be accessed by
multiple theads" in psm.exe / libdefault.so / LibPSM.suprx,

this spelling error is intentional please do not fix it,

+++ subprojects/LibPSM/Sce/Pss/Core/Graphics/ShaderProgram.cpp
@@ -313,12 +313,12 @@ namespace Sce::Pss::Core::Graphics {
 		}
 
 		GL_CALL(glBindAttribLocation(this->GLReference, index, name.c_str()));
-		if (attributeBindings.size() <= index) attributeBindings.resize(index + 1);
+		if (attributeBindings.size() <= static_cast<size_t>(index)) attributeBindings.resize(index + 1);
 		attributeBindings[index] = name;
 		return PSM_ERROR_NO_ERROR;
 	}
 	std::string ShaderProgram::GetAttributeBinding(int index) const {
-		if(attributeBindings.size() <= index) return "";
+		if(attributeBindings.size() <= static_cast<size_t>(index)) return "";
 		return attributeBindings[index];
     }
